requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
twilio==9.0.5
python-dotenv==1.0.1
